removing old component
