--- src/screen/backoffice/backoffice.module.css
@@ -1,86 +1,0 @@
-:root {
-  --bg-header: #fff;
-  --bg-main: #edf2f7;
-  --bg-card: #fff;
-  --color-card-title: #007bff;
-  --bg-card-button: #319795;
-  --color-card-button: #fff;
-}
-
-.backoffice {
-  background-color: var(--bg-main);
-  min-height: 100vh;
-}
-.backoffice__header {
-  height: 60px;
-  background-color: var(--bg-header);
-  border-bottom: 0.0625rem solid #ccc;
-  margin-bottom: 2.5rem;
-}
-.backoffice__main {
-  width: 90%;
-  max-width: 68.75rem;
-  margin: 0 auto;
-}
-.backoffice__ul {
-  display: grid;
-  column-gap: 4.5rem;
-  row-gap: 0.75rem;
-  grid-auto-rows: 10rem;
-  justify-content: center;
-  grid-template-columns: repeat(
-    auto-fill,
-    minmax(min(100%, 10.3125rem), 10.3125rem)
-  );
-}
-@media screen and (min-width: 75rem) {
-  .backoffice__ul {
-    column-gap: 9rem;
-  }
-}
-.backoffice__li {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border: 0.0625rem solid rgba(255, 255, 255, 0.2);
-  padding-bottom: 0.5rem;
-  padding-top: 1rem;
-  background-color: var(--bg-card);
-  border-radius: 0.5rem;
-  box-shadow: 0rem 0.125rem 0.1875rem 0rem rgba(0, 0, 0, 0.1);
-  -webkit-box-shadow: 0rem 0.125rem 0.1875rem 0rem rgba(0, 0, 0, 0.1);
-  -moz-box-shadow: 0rem 0.125rem 0.1875rem 0rem rgba(0, 0, 0, 0.1);
-}
-.backoffice__h2 {
-  margin-bottom: 0.5rem;
-  color: var(--color-card-title);
-}
-.backoffice__image-container {
-  width: 4rem;
-  height: 4rem;
-  border-radius: 50%;
-  background-color: rgba(0, 0, 0, 0.03);
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.backoffice__image {
-  height: 3.375rem;
-  width: 3.375rem;
-}
-.backoffice__button {
-  margin-top: auto;
-  background-color: var(--bg-card-button);
-  color: var(--color-card-button);
-  height: 1.5rem;
-  width: 1.5rem;
-  border-radius: 0.25rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 0.75rem;
-}
-.backoffice__button:hover {
-  cursor: pointer;
-}
